pyproject.toml
[project]
name = "iperf-lattice"
version = "0.1.0"
description = "iperf3 mesh orchestrator — discover, matrix, and flood modes for cluster network testing"
requires-python = ">=3.11"
dependencies = [
    "asyncssh>=2.17",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.scripts]
iperf-lattice = "iperf_lattice:main"
iperf-lattice-discover = "iperf_lattice_discover:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["iperf_lattice", "iperf_lattice_discover"]

[tool.ruff.format]
quote-style = "double"
